djangoapp/requirements.txt
Django>=4.2.3,<4.3
Django-stubs>=5.2.8,<5.3
psycopg2-binary>=2.9.6,<2.10
Pillow>=9.5.0,<9.6
python-dotenv>=1.0.0,<1.1
django-axes>=6.0.5,<6.1
django-crispy-forms>=1.13.0,<1.14
qrcode>=7.4.2,<7.5
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3.0,<5.4
exponent_server_sdk>=2.1.0,<2.2
celery[redis]>=5.3.0,<5.4
redis>=4.5.4,<4.6

